Cycle in LinkedList
The age old problem of detecting a cycle in a LinkedList. Logic used -
First use the slow and fast pointer to verify if a cycle exists.

If the linked list does have a cycle, then move the slow pointer to head
of the list and then move both the slow and fast pointer one at a time
till they co-incide. That would be the point where the loop starts.